pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "audio-transcriber"
version = "1.0.0"
description = "MP4動画から話者分離付き文字起こしを自動生成するツール"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.9"
authors = [
    {name = "Audio Transcriber", email = "example@example.com"},
]

dependencies = [
    "torch>=2.0.0",
    "torchaudio>=2.0.0",
    "transformers>=4.30.0",
    "pyannote.audio>=3.1.0",
    "moviepy>=1.0.3",
    "numpy>=1.24.0",
]

[project.optional-dependencies]
web = [
    "streamlit>=1.28.0",
]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
]

[tool.uv]
dev-dependencies = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
]

# Hatchlingがファイルを見つけられるように明示的に指定
[tool.hatch.build.targets.wheel]
packages = ["."]
exclude = [
    "/.git",
    "/.venv",
    "__pycache__",
    "*.pyc",
    ".DS_Store",
]

[tool.black]
line-length = 88
target-version = ['py39', 'py310', 'py311', 'py312']

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88
